Initial Commit from project back in 1996-1997 era
Before embracing Linux I had a growing dislike of Microsoft and set out back in high school to create a replacement.  This is the result of those efforts.    The assembler was quite advanced in it's development, based heavily on nasm and the operating system would boot off a floppy disk and boot entering protected mode, but not much else was completed as I recall.